Fix the failing `ingestion` tests (#7)
# PULL REQUEST

## Summary

I added a FHIR bundle that was the result of our FHIR conversion process
derrived from a valid eICR/RR pair. The bundle used can be found
[here](https://github.com/CDCgov/dibbs-product-demos/blob/main/ecr-parser/data/fhir-conversion/eICR-TC-COVID-LabPos_20210412-bundle.json)
on the `dibbs-product-demos` repo. Since the test is asserting the known
length of resources in the original bundle I adjusted the number in the
test.

## Related Issue

Fixes #3 
- #3 

## Acceptance Criteria

All 57 tests are passing locally.